Try to make go imports work as separated cmd.
goimports is useful tools, but if it is triggered in on-save, it is a little bit annoying when coding golang.
So, I duplicate go_fmt with new go_imports, in order to give user a chance to set up below keybinding to run goimports mannually.

So, the better user experience is , we set go_fmt while on-save, and use go_imports when we need it.

```
{
        "keys": ["super+.", "super+f"],
        "command": "gs_imports",
        "context": [{ "key": "selector", "operator": "equal", "operand": "source.go" }]
    },
```